Moved several classes into Examples directory / namespace
